# api/openapi.yaml
openapi: 3.0.3
info:
  title: Hotel Reviews Microservice API
  description: |
    A comprehensive microservice for managing hotel reviews, ratings, and related data.
    
    ## Features
    - Hotel and provider management
    - Review creation, retrieval, and management
    - Advanced search and filtering capabilities
    - File upload and batch processing
    - User authentication and authorization
    - Real-time metrics and monitoring
    
    ## Authentication
    This API uses JWT Bearer token authentication. Obtain a token by calling the `/auth/login` endpoint.
    
    ## Rate Limiting
    API requests are rate-limited. Check the `X-RateLimit-*` headers in responses for current limits.
    
    ## Versioning
    The API uses URL path versioning (e.g., `/api/v1/`). Current version is v1.
  version: 1.0.0
  contact:
    name: Hotel Reviews API Support
    email: support@hotelreviews.com
    url: https://docs.hotelreviews.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://hotelreviews.com/terms
servers:
  - url: https://api.hotelreviews.com/api/v1
    description: Production server
  - url: https://staging-api.hotelreviews.com/api/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Development server
paths:
...
  /reviews:
    get:
      tags:
        - Reviews
      summary: List reviews
      description: Retrieve a paginated list of reviews with optional filtering
      operationId: listReviews
      security:
        - BearerAuth: []
      parameters:
        - name: hotel_id
          in: query
          description: Filter reviews by hotel ID
          schema:
            type: string
            format: uuid
        - name: provider_id
          in: query
          description: Filter reviews by provider ID
          schema:
            type: string
            format: uuid
        - name: rating_min
          in: query
          description: Minimum rating filter
          schema:
            type: number
            minimum: 1
            maximum: 5
        - name: rating_max
          in: query
          description: Maximum rating filter
          schema:
            type: number
            minimum: 1
            maximum: 5
        - name: language
          in: query
          description: Filter by review language
          schema:
            type: string
            pattern: '^[a-z]{2}$'
            example: "en"
        - name: from_date
          in: query
          description: Filter reviews from this date
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          description: Filter reviews to this date
          schema:
            type: string
            format: date
        - name: limit
          in: query
          description: Number of reviews to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of reviews to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [rating, review_date, created_at, helpful_votes]
            default: created_at
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'
              examples:
                review_list:
                  summary: List of reviews
                  value:
                    success: true
                    data:
                      reviews:
                        - id: "550e8400-e29b-41d4-a716-446655440000"
                          hotel_id: "123e4567-e89b-12d3-a456-426614174000"
                          provider_id: "987fcdeb-51a2-4567-8901-23456789abcd"
                          rating: 4.5
                          title: "Great stay!"
                          comment: "Had a wonderful time at this hotel."
                          review_date: "2024-01-10T00:00:00Z"
                          language: "en"
                          is_verified: true
                          helpful_votes: 12
                          total_votes: 15
                          created_at: "2024-01-11T10:30:00Z"
                      pagination:
                        total: 156
                        limit: 20
                        offset: 0
                        has_next: true
                        has_previous: false
                    timestamp: "2024-01-15T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Reviews
      summary: Create a new review
      description: Create a new hotel review
      operationId: createReview
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
            examples:
              basic_review:
                summary: Basic review
                value:
                  hotel_id: "123e4567-e89b-12d3-a456-426614174000"
                  reviewer_info_id: "456e7890-e89b-12d3-a456-426614174001"
                  rating: 4.5
                  title: "Great stay!"
                  comment: "Had a wonderful time at this hotel. Staff was friendly and rooms were clean."
                  review_date: "2024-01-10T00:00:00Z"
                  language: "en"
              detailed_review:
                summary: Detailed review with all ratings
                value:
                  hotel_id: "123e4567-e89b-12d3-a456-426614174000"
                  reviewer_info_id: "456e7890-e89b-12d3-a456-426614174001"
                  rating: 4.2
                  title: "Good hotel with excellent service"
                  comment: "Overall a good experience. The location was perfect and staff very helpful."
                  review_date: "2024-01-10T00:00:00Z"
                  language: "en"
                  service_rating: 5.0
                  cleanliness_rating: 4.0
                  location_rating: 5.0
                  value_rating: 3.5
                  comfort_rating: 4.0
                  facilities_rating: 4.0
                  trip_type: "business"
                  room_type: "standard"
      responses:
        '201':
          description: Review created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /hotels:
    get:
      tags:
        - Hotels
      summary: List hotels
      description: Retrieve a paginated list of hotels with optional filtering
      operationId: listHotels
      security:
        - BearerAuth: []
      parameters:
        - name: name
          in: query
          description: Filter by hotel name (partial match)
          schema:
            type: string
            maxLength: 255
        - name: city
          in: query
          description: Filter by city
          schema:
            type: string
            maxLength: 100
        - name: country
          in: query
          description: Filter by country
          schema:
            type: string
            maxLength: 100
        - name: star_rating
          in: query
          description: Filter by star rating
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: min_rating
          in: query
          description: Minimum average review rating
          schema:
            type: number
            minimum: 1
            maximum: 5
        - name: limit
          in: query
          description: Number of hotels to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of hotels to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Hotels retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Hotels
      summary: Create a new hotel
      description: Create a new hotel (admin only)
      operationId: createHotel
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHotelRequest'
      responses:
        '201':
          description: Hotel created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the authentication endpoints
  schemas:
...
    CreateReviewRequest:
      type: object
      required:
        - hotel_id
        - reviewer_info_id
        - rating
        - title
        - comment
        - review_date
        - language
      properties:
        hotel_id:
          type: string
          format: uuid
          description: Hotel ID
        reviewer_info_id:
          type: string
          format: uuid
          description: Reviewer information ID
        external_id:
          type: string
          maxLength: 255
          nullable: true
          description: External system review ID
        rating:
          type: number
          minimum: 1.0
          maximum: 5.0
          multipleOf: 0.1
          description: Overall rating
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Review title
        comment:
          type: string
          minLength: 1
          description: Review text content
        review_date:
          type: string
          format: date-time
          description: Date when review was written
        stay_date:
          type: string
          format: date
          nullable: true
          description: Date of hotel stay
        trip_type:
          type: string
          maxLength: 50
          nullable: true
          description: Type of trip
        room_type:
          type: string
          maxLength: 100
          nullable: true
          description: Room type
        language:
          type: string
          pattern: '^[a-z]{2}$'
          description: Review language code
        service_rating:
          type: number
          minimum: 1.0
          maximum: 5.0
          nullable: true
          description: Service rating
        cleanliness_rating:
          type: number
          minimum: 1.0
          maximum: 5.0
          nullable: true
          description: Cleanliness rating
        location_rating:
          type: number
          minimum: 1.0
          maximum: 5.0
          nullable: true
          description: Location rating
        value_rating:
          type: number
          minimum: 1.0
          maximum: 5.0
          nullable: true
          description: Value rating
        comfort_rating:
          type: number
          minimum: 1.0
          maximum: 5.0
          nullable: true
          description: Comfort rating
        facilities_rating:
          type: number
          minimum: 1.0
          maximum: 5.0
          nullable: true
          description: Facilities rating
...
    ReviewResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Review'
    ReviewListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                reviews:
                  type: array
                  items:
                    $ref: '#/components/schemas/Review'
                pagination:
                  $ref: '#/components/schemas/Pagination'
...
    CreateHotelRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Hotel name
        address:
          type: string
          description: Hotel address
        city:
          type: string
          maxLength: 100
          description: City
        country:
          type: string
          maxLength: 100
          description: Country
        postal_code:
          type: string
          maxLength: 20
          description: Postal code
        phone:
          type: string
          maxLength: 20
          description: Phone number
        email:
          type: string
          format: email
          maxLength: 255
          description: Email address
        star_rating:
          type: integer
          minimum: 1
          maximum: 5
          description: Official star rating
        description:
          type: string
          description: Hotel description
        amenities:
          type: array
          items:
            type: string
          description: List of amenities
        latitude:
          type: number
          minimum: -90
          maximum: 90
          description: Latitude coordinate
        longitude:
          type: number
          minimum: -180
          maximum: 180
          description: Longitude coordinate
...
    HotelResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Hotel'
    HotelListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                hotels:
                  type: array
                  items:
                    $ref: '#/components/schemas/Hotel'
                pagination:
                  $ref: '#/components/schemas/Pagination'

    # Provider Schemas
...
    BaseResponse:
      type: object
      required:
        - success
        - timestamp
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
        trace_id:
          type: string
          nullable: true
          description: Request trace ID for debugging
...
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - error
            - error_code
          properties:
            success:
              type: boolean
              enum: [false]
            error:
              type: string
              description: Error message
            error_code:
              type: string
              description: Machine-readable error code
            details:
              type: object
              nullable: true
              description: Additional error details
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            details:
              type: object
              properties:
                validation_errors:
                  type: array
                  items:
                    type: object
                    properties:
                      field:
                        type: string
                        description: Field that failed validation
                      message:
                        type: string
                        description: Validation error message
                      code:
                        type: string
                        description: Validation error code

    # Utility Schemas
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
          description: Maximum items per page
        offset:
          type: integer
          description: Number of items skipped
        has_next:
          type: boolean
          description: Whether there are more pages
        has_previous:
          type: boolean
          description: Whether there are previous pages
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_parameter:
              summary: Invalid parameter
              value:
                success: false
                error: "Invalid parameter value"
                error_code: "INVALID_PARAMETER"
                timestamp: "2024-01-15T10:30:00Z"
                trace_id: "abc123def456"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing_token:
              summary: Missing authentication token
              value:
                success: false
                error: "Authentication required"
                error_code: "UNAUTHORIZED"
                timestamp: "2024-01-15T10:30:00Z"
                trace_id: "abc123def456"
            invalid_token:
              summary: Invalid authentication token
              value:
                success: false
                error: "Invalid or expired token"
                error_code: "INVALID_TOKEN"
                timestamp: "2024-01-15T10:30:00Z"
                trace_id: "abc123def456"
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            insufficient_permissions:
              summary: Insufficient permissions
              value:
                success: false
                error: "Insufficient permissions for this operation"
                error_code: "FORBIDDEN"
                timestamp: "2024-01-15T10:30:00Z"
                trace_id: "abc123def456"
...
    ValidationError:
      description: Validation error - invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          examples:
            validation_failure:
              summary: Validation errors
              value:
                success: false
                error: "Request validation failed"
                error_code: "VALIDATION_ERROR"
                details:
                  validation_errors:
                    - field: "email"
                      message: "Invalid email format"
                      code: "INVALID_FORMAT"
                    - field: "rating"
                      message: "Rating must be between 1 and 5"
                      code: "OUT_OF_RANGE"
                timestamp: "2024-01-15T10:30:00Z"
                trace_id: "abc123def456"
...